Fix JSDoc comments for StateLoader

+++ .eslintrc.js
@@ -7,7 +7,7 @@ module.exports = {
     // allow debugger during development
     "no-debugger": process.env.NODE_ENV === "production" ? 2 : 0,
     // VS Code / MSVS autoformat rules
-    quotes: [2, "double"],
+    quotes: [2, "single"],
     semi: [2, "always"],
     "space-before-function-paren": [2, {
       "anonymous": "always",
